WIP: Задание 2, класс с препятствием.Проверка
