+++ Char/Char.csproj
@@ -43,8 +43,8 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="Charactor.cs" />
-    <Compile Include="Let.cs" />
+    <Compile Include="ClassCharactor.cs" />
+    <Compile Include="ClassLet.cs" />
     <Compile Include="Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
   </ItemGroup>
